move price requests to API
